feat(global): add reusable BackToTopButton component and integrate in all pages
- Extracted BackToTopButton from individual pages into a standalone component.
- Integrated BackToTopButton across all main pages for consistent UX.
- Ensured smooth scroll and responsive display with Framer Motion.
