+++ frontend/src/pages/about/about.jsx
@@ -12,6 +12,7 @@ import {
   FaUsers,
 } from 'react-icons/fa';
 import axios from 'axios';
+import BackToTopButton from '@/components/utils/backToTopButton';
 
 export default function About() {
   const BACKEND_URL = import.meta.env.VITE_URL_BACKEND;
@@ -149,12 +150,12 @@ export default function About() {
         keywords='TerraQuake API, earthquake monitoring, seismic data, earthquake detection API, real-time earthquake data, disaster prevention, seismic analysis'
       />
       {/* SEO Stuff */}
-      
+
       {/* Background Gradient/Mesh (for a classy, dark theme) */}
-        <div className='absolute inset-0 z-0'>
-          <div className='absolute top-0 left-0 w-80 h-80 bg-purple-500 rounded-full mix-blend-multiply filter blur-3xl opacity-30 animate-blob' />
-          <div className='absolute bottom-10 right-10 w-96 h-96 bg-pink-500 rounded-full mix-blend-multiply filter blur-3xl opacity-30 animate-blob animation-delay-2000' />
-        </div>
+      <div className='absolute inset-0 z-0'>
+        <div className='absolute top-0 left-0 w-80 h-80 bg-purple-500 rounded-full mix-blend-multiply filter blur-3xl opacity-30 animate-blob' />
+        <div className='absolute bottom-10 right-10 w-96 h-96 bg-pink-500 rounded-full mix-blend-multiply filter blur-3xl opacity-30 animate-blob animation-delay-2000' />
+      </div>
 
       <section className='relative z-30 w-full min-h-screen px-6 py-20 overflow-hidden'>
         {/* Header Section */}
@@ -255,6 +256,8 @@ export default function About() {
           ))}
         </div>
       </section>
+      {/* Floating Back-to-Top Button Component */}
+      <BackToTopButton />
     </>
   );
 }
