+++ frontend/src/pages/profile/profile.jsx
@@ -5,6 +5,7 @@ import Swal from 'sweetalert2';
 import { useNavigate } from 'react-router-dom';
 import EditProfile from './editProfile';
 import DeleteProfile from './deleteProfile';
+import BackToTopButton from '@/components/utils/backToTopButton';
 
 export default function Profile() {
   const { userLogin, isLoggedIn, setIsLoggedIn, setUserLogin } =
@@ -187,6 +188,8 @@ export default function Profile() {
           </div>
         )}
       </section>
+      {/* Floating Back-to-Top Button Component */}
+      <BackToTopButton />
     </>
   );
 }
